ui: Use (time) scale instead of currentWindow for metrics page graph
This commit modifies metrics page graphs to use the `scale` property on the time
window state, instead of `currentWindow`. It does so by setting the `windowEnd`
property on `scale` when a user drags-to-zoom the graph, and reading from
`scale` to query for data in the provider.

This commit does not regress #70594 which initially prompted using
`currentWindow`. Note that at the code state when the #70594 bug was fixed, the
`windowEnd` property was not set (this commit changes `linegraph/index.tsx` to
set it). Even when `windowEnd` is not set, reading the start and end times from
`scale` did not recreate the drag-to-zoom granularity bug. However, reading
from `scale` without setting `windowEnd` does cause a separate major bug in
that the timescale endpoint is always queried with `now` as an end time instead
of the time that the user selected and is displayed.

In light of this potential major bug, the videos are taken with an additional
`console.log()` (that has not been committed) in the `sendRequestBatch` of
`redux/metrics.ts`. It logs what end time is being used in the request, to
verify correctness.

This is done in preparation of de-duplicating this information about start in
and time by removing `currentWindow`, in preparation for #72128.

Release note: None