ui: Add timescale component to statement details and transaction details pages
Addresses #74512

This commit adds the time selection component to the Statement Details
(Overview and Explain Plan tabs) and Transaction Details pages.

Release note (ui): The time selection component previously on the Statements and
Transactions Pages has been added to the Statement Details Overview and Explain
Plan tabs, and the Transaction Details page. Changing the selected time in any
of these locations will change the time selected for all of them.