ui: Add time picker to Transaction Details and Statement Details Overview pages
Addresses #74512

This commit adds the time selection component to the Statement Details
(Overview and Explain Plan tabs) and Transaction Details pages.

If staying on the same Statement/Transaction details page and switching to a
time that there is no data for, the Overview and Explain Plan tabs on the
Statement Details page, and the Transaction details page will continue to show
the statement/transaction text. The Diagnostics tab on the Statement Details
page will also continue to work as usual. However, these features are not
available if landing directly on a Statement/Transactions Details page for a
time frame with no data, without having previously switched from a time frame
that did have data for the same statement/transaction.

This commit also fixes Transaction Details page stories that were previously out
of date and broken. The Statement Details Explain Plan tab story was already
previously broken and is not fixed here.

Release note (ui): The time selection component previously on the Statements and
Transactions Pages has been added to the Statement Details Overview and Explain
Plan tabs, and the Transaction Details page.