ui: Add time frame shown to Jobs Page table
Fixes #71563

This commit adds a message in the Jobs Page table indicating the oldest time
(in UTC) that jobs are shown for. This oldest time is calculated as `now - the
duration of the jobs.retention_time cluster setting`.

Release note (ui): The Jobs Page now shows the oldest time (in UTC) that jobs
are shown for.